refactor type definition
