+++ types/index.d.ts
@@ -1,4 +1,4 @@
 import './vue'
-import { VueApollo } from './vue-apollo';
+import { VueApollo } from './vue-apollo'
 
-export default VueApollo;
+export default VueApollo
